backend/pyproject.toml
[project]
name = "saas-backend"
version = "0.1.0"
description = "Multi-tenant SaaS backend"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "bcrypt>=4.0.0",
    "pydantic-settings>=2.7.0",
    "pyjwt>=2.9.0",
    "email-validator>=2.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
